requirements.txt
fastapi
uvicorn
jinja2
requests
netmiko
python-multipart
python-dotenv